+++ plugins/org.preesm.codegen/src/org/preesm/codegen/xtend/task/CodegenEngine.java
@@ -97,6 +97,9 @@ public class CodegenEngine {
   /** The codegen path. */
   private final String codegenPath;
 
+  /** The Apollo flag. */
+  private boolean apolloEnabled = false;
+
   /** The code blocks. */
   private final Collection<Block> codeBlocks;
 
@@ -305,6 +308,9 @@ public class CodegenEngine {
       // initialize printer engine
       printer.setEngine(this);
 
+      // set up apollo intra-actor optimization
+      printer.setApolloEnabled(this.apolloEnabled);
+
       // Do the pre-processing
       printer.preProcessing(printerAndBlocks.getValue(), this.codeBlocks);
       this.realPrinters.put(printerAndBlocks.getKey(), printer);
@@ -379,5 +385,14 @@ public class CodegenEngine {
         PreesmLogger.getLogger().log(Level.FINE, msg);
     }
 
+  }
+
+  /**
+   * 
+   * @param apolloFlag
+   *          Enable the use of Apollo for intra-actor optimization
+   */
+  public void registerApollo(String apolloFlag) {
+    this.apolloEnabled = "true".equalsIgnoreCase(apolloFlag);
   }
 }